move images from general assets to dedicated home sections
