+++ src/routes/home/sections/welcome/assets/pizza-one-parallax.png
