+++ src/routes/home/sections/welcome/assets/pizza-two-parallax.png
