+++ src/routes/home/sections/welcome/assets/section-one-700.webp
